+++ CHANGELOG.md
@@ -15,6 +15,7 @@
   - Add analysis class for readout calibration. (!474)
   - Fix units of Rabi analysis amplitudes. (!511)
   - Remove an `xarray` `FutureWarning` in `to_gridded_dataset` when retrieving the dataset dimension names. (!510)
+  - Remove an `xarray` `DeprecationWarning` about `argmin` and `argmax` in the `RabiAnalysis`. (!513)
   - Remove a `matplotlib` `MatplotlibDeprecationWarning` in `Basic2DAnalysis` when plotting a heatmap. (!510)
 - GitLab
   - Make general Merge Request template the default. (!507)